rework Kconfig. remove mmu map 0xff000000 usage. remove hardcoded mmu mapping before start_kernel. rework debug config. add serial output debug during assembly part. use our own linker script for custom kernel offset. use ioremap only. improve speed on aout generation. cleanup eth driver. add debug config to scsi. cleanup scsi. improve framebuffer. add gray vga16 linux next logo for mono.

+++ arch/m68k/tools/next/next-disklabel.c
@@ -188,7 +188,7 @@ int write_dl(struct next68k_disklabel *dl, FILE *diskf) {
 	struct stat buffer;
 	fstat(disk_fd, &buffer);
 	off_t disk_size = buffer.st_size;
-	printf("Disk size is %ld\n", disk_size);
+	printf("Disk size is %jd\n", (intmax_t)disk_size);
 
 	bzero(dl->cd_label, NEXT68K_LABEL_CPULBLLEN);
 	strncpy(dl->cd_label, "New Disk", NEXT68K_LABEL_CPULBLLEN-1);
